Fixed swagger error and database
