+++ budget-manager-server/src/BudgetManager.Application/Transactions/CreateTransaction/CreateTransactionRequest.cs
@@ -1,4 +1,4 @@
 ﻿using BudgetManager.Domain.Enums;
 
-namespace BudgetManager.Application.Users.CreateTransaction;
+namespace BudgetManager.Application.Transactions.CreateTransaction;
 public record CreateTransactionRequest(Guid CategoryId, TransactionType TransactionType, DateTime TransactionDate, decimal Amount, string? Description);
