+++ budget-manager-server/src/BudgetManager.Application/Transactions/CreateTransaction/CreateTransactionResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace BudgetManager.Application.Transactions.CreateTransaction;
+
+public record CreateTransactionResponse(Guid TransactionId);
