--- budget-manager-server/src/BudgetManager.Application/Users/CreateTransaction/CreateTransactionResponse.cs
@@ -1,3 +1,0 @@
-﻿namespace BudgetManager.Application.Users.CreateTransaction;
-
-public record CreateTransactionResponse(Guid TransactionId);
