+++ budget-manager-server/src/BudgetManager.Infrastructure/BudgetManager.Infrastructure.csproj
@@ -11,6 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Data\Migrations\" />
     <Folder Include="Settings\" />
   </ItemGroup>
